reworked button callbacks to use lambdas than run outside of the ECS update loop.

+++ Engine/src/engine/math/Vec2.cpp
@@ -11,8 +11,8 @@ namespace se::math
     DEFINE_SPARK_CLASS_END()
 
     Vec2::Vec2(const IntVec2 &v)
-        : x(v.x)
-        , y(v.y)
+        : x(static_cast<float>(v.x))
+        , y(static_cast<float>(v.y))
     {
 
     }
